+++ fe/src/components/coupon/CouponGroupCard.tsx
@@ -6,7 +6,7 @@ import TextInputGroup from './inputGroup/TextInputGroup';
 import { useInput } from '@hooks/useInput';
 import { useCoupon } from '@components/coupon/useCoupons';
 import SelectInputGroup from './inputGroup/SelectInputGroup';
-import CustomButton from '@components/common/button/CustomButton';
+import CustomButton from '@components/common/CustomButton/CustomButton';
 import { Icon } from '@components/common/Icon/Icon';
 import { CouponGroupType, IsRandom } from './CouponType';
